Rename Mainnet -> Ethereum all over the codebase.
- Rename Mainnet to Ethereum becuase as we branch out to other L1's
  (such as Avalanche), calling it Mainnet makes less sense and may be
confusing for non Ethereum natives.
