+++ src/config/index.ts
@@ -2,7 +2,7 @@ import { Network, Configs } from '../types'
 import arbitrum from './arbitrum'
 import gnosis from './gnosis'
 import goerli from './goerli'
-import mainnet from './mainnet'
+import ethereum from './ethereum'
 import optimism from './optimism'
 import polygon from './polygon'
 import zkevm from './zkevm'
@@ -11,7 +11,7 @@ const config: Configs = {
   [Network.Arbitrum]: arbitrum,
   [Network.Gnosis]: gnosis,
   [Network.Goerli]: goerli,
-  [Network.Mainnet]: mainnet,
+  [Network.Ethereum]: ethereum,
   [Network.Optimism]: optimism,
   [Network.Polygon]: polygon,
   [Network.Zkevm]: zkevm,
