+++ src/lib/templates/tokenlist/overwrites.ts
@@ -1,10 +1,11 @@
 import { Network, OverwritesForList } from '../../types'
 
 export const overwrites: OverwritesForList = {
-  [Network.Mainnet]: {},
+  [Network.Ethereum]: {},
   [Network.Goerli]: {},
   [Network.Polygon]: {},
   [Network.Arbitrum]: {},
   [Network.Optimism]: {},
   [Network.Gnosis]: {},
+  [Network.Zkevm]: {},
 }
