+++ src/lib/templates/tokenlist/tokens.ts
@@ -1,10 +1,11 @@
 import { Network, TokensForList } from '../../types'
 
 export const tokens: TokensForList = {
-  [Network.Mainnet]: [],
+  [Network.Ethereum]: [],
   [Network.Goerli]: [],
   [Network.Polygon]: [],
   [Network.Arbitrum]: [],
   [Network.Optimism]: [],
   [Network.Gnosis]: [],
+  [Network.Zkevm]: [],
 }
