+++ src/tokenlists/listed-old/overwrites.ts
@@ -2,7 +2,7 @@
 import { Network, OverwritesForList } from '../../types'
 
 export const overwrites: OverwritesForList = {
-  [Network.Mainnet]: {
+  [Network.Ethereum]: {
     '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2': {
       name: 'Wrapped Ether',
     },
