+++ src/tokenlists/listed-old/tokens.ts
@@ -1,7 +1,7 @@
 import { Network, TokensForList } from '../../types'
 
 export const tokens: Partial<TokensForList> = {
-  [Network.Mainnet]: [
+  [Network.Ethereum]: [
     '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2',
     '0xba100000625a3754423978a60c9317c58a424e3D',
     '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48',
